code gardening

+++ tracevis/traceMisc.cpp
@@ -117,7 +117,7 @@ bool obtainMutex(HANDLE mutex, char *errorLocation, int waitTime)
 	DWORD waitresult = WaitForSingleObject(mutex, waitTime);
 	if (waitresult == WAIT_TIMEOUT) {
 		if (errorLocation)
-			printf("WARNING! Mutex %x wait expired at %s ERROR!\n", mutex, errorLocation);
+			printf("WARNING! Mutex %x wait expired at %s ERROR!\n", (unsigned int)mutex, errorLocation);
 		return false;
 	}
 	//if (errorLocation)
